REVERTME: Support 64-bit address
The VXD driver cast 64-bit address ont 32-bit. This is
a temporary fix until we get a new delivery from
3rd party supplier. This delivery is tracked in: APDEV-643.

Issue: APDEV-274

Depends-Change-Id: I8e496a6001c13a00c2e6170cdb9cc538ff0ad0b6
Depends-Change-Id: I56f71f1f4ad8f0e25c59f6a6e4c73a535c47e220
Depends-Change-Id: Ic8d311bf724a8726d2ff1fa27acbd0a94bfc4a30
Depends-Change-Id: I77704daf04f967936c09210f010bd59deddc54b7
Depends-Change-Id: Icc73bf55181aa1415c38d3d48b57dc4abd937304
Depends-Change-Id: Id80c269f963210785b86e1ebd05015b95379f514

+++ video/decode/psb_msvdx_reg.h
@@ -87,7 +87,7 @@ static inline uint32_t PSB_RMSVDX32(uint32_t _offs)
 #define MTX_PC		MTX_INTERNAL_REG(0, 5)
 
 #define MEMIO_READ_FIELD(vpMem, field)						\
-	((uint32_t)(((*((field##_TYPE*)(((uint32_t)vpMem) + field##_OFFSET))) \
+	((uint64_t)(((*((field##_TYPE*)(((uint64_t)vpMem) + field##_OFFSET))) \
 			& field##_MASK) >> field##_SHIFT))			\
 
 #define MEMIO_WRITE_FIELD(vpMem, field, value)					\
